api1.0/requirements.txt
Django>=4.2
djangorestframework>=3.14
qrcode>=7.3
